requirements.txt
aiohttp
Pillow==10.0.0
pydantic
aiofiles
